requirements.txt
Flask==1.1.1
gunicorn==19.9.0
pip==20.2.3 
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
tensorflow-cpu
matplotlib>=1.4.3
pandas>=0.19
seaborn>=0.9.1
jsonschema==3.2.0
